% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guimidrangeMCP(tcltk).R
\name{guimidrangeMCP}
\alias{guimidrangeMCP}
\title{Graphical User Interface for midrangeMCP function}
\usage{
guimidrangeMCP(gui = TRUE)
}
\arguments{
\item{gui}{Logical argument, \code{TRUE} or \code{FALSE}. The default is \code{TRUE}}
}
\value{
\code{guimidrangeMCP} presents a GUI for the results of the four multiple
      comparison procedures MGM, MGR, SNKM and TM tests. In addition, the GUI
      returns a graph of the results, as well as the export of these results
      to three types of file extension and latex code.
}
\description{
\code{guimidrangeMCP} A Graphical User Interface (GUI) for
    function that returns the MGM, MGR, SNKM and TM tests
}
\examples{
# Loading package
library(midrangeMCP)
if (interactive()) {
  guimidrangeMCP(gui = FALSE)
}

}
