% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRtest.R
\name{MRtest}
\alias{MRtest}
\title{Multiple comparison procedures to the means of a factor using
the studentized range and midrange distributions.}
\usage{
MRtest(
  y,
  trt = NULL,
  dferror = NULL,
  mserror = NULL,
  replication = NULL,
  alpha = 0.05,
  main = NULL,
  MCP = "all",
  ismean = FALSE
)
}
\arguments{
\item{y}{Model (aov or lm), numeric vector containing the
response variable or the mean of the treatments.}

\item{trt}{Constant (y = model) or a vector containing
the treatments.}

\item{dferror}{Degrees of freedom of the Mean Square Error.}

\item{mserror}{Mean Square Error.}

\item{replication}{Number de repetitions of the treatments in the experiment.
For unbalanced data should be informed the harmonic mean of repetitions.
This argument should be informed only if \code{ismean = TRUE}.}

\item{alpha}{Significant level. The default is \code{alpha = 0.05}.}

\item{main}{Title of the analysis.}

\item{MCP}{Allows choosing the multiple comparison test;
the \emph{defaut} is "all". This option will go perform all tests.
However, the options are: the Skott-Knott midrange test ("MGM"),
the Skott-Knott Range test ("MGR"), the Student-Newman-Keuls
midrange test ("SNKM") and the Tukey midrange test ("TM").}

\item{ismean}{Logic. If \code{FALSE} (default), the \code{y} argument represents
a model (aov or lm) or a numeric vector containing the response variable. If \code{TRUE} the \code{y} argument represents the
mean of treatments.}
}
\value{
\code{MRtest} returns the print of a list of results. First,
    the summary of \code{y}. Second, the statistics
    of the test chosen. And finally, the  mean group results for each test.
    If \code{MRtest} function is stored
    in an object, the results will be printed and
    also stored in the object.
}
\description{
\code{MRtest} applies the Skott-Knott midrange, Skott-Knott range,
    Student-Newman-Keuls midrange and Tukey midrange tests. These are new
    tests for multiple comparisons proposed by the authors (2015), that are in publication
    fase.
}
\details{
The \code{MCP} argument allows you to choose various tests
    of multiple comparisons at once. For example,
    \code{MCP = c("MGM", "MGR")}, and so on.
}
\examples{
# Simulated data (completely randomized design)

# Response variable
rv <- c(100.08, 105.66, 97.64, 100.11, 102.60, 121.29, 100.80,
        99.11, 104.43, 122.18, 119.49, 124.37, 123.19, 134.16,
        125.67, 128.88, 148.07, 134.27, 151.53, 127.31)

# Treatments
treat <- factor(rep(LETTERS[1:5], each = 4))

# Anova
res     <- anova(aov(rv~treat))
DFerror <- res$Df[2]
MSerror <- res$`Mean Sq`[2]

# Loading the midrangeMCP package
library(midrangeMCP)

# applying the tests
results <- MRtest(y = rv,
                  trt = treat,
                  dferror = DFerror,
                  mserror = MSerror,
                  alpha = 0.05,
                  main = "Multiple Comparison Procedure: MGM test",
                  MCP = c("MGM"))

# Other option for the MCP argument is "all". All tests are used.

results$Groups     # Results of the tests
results$Statistics # Main arguments of the tests
results$Summary    # Summary of the response variable

# Using the y argument as aov or lm model
res  <- aov(rv~treat)

MRtest(y = res, trt = "treat", alpha = 0.05,
       main = "Multiple Comparison Procedure: MGM test",
       MCP = c("MGM"))

# For unbalanced data: It will be used the harmonic mean of
#                       the number of experiment replicates

# Using the previous example
rv <- rv[-1]
treat <- treat[-1]

res  <- lm(rv~treat) # Linear model

# Multiple comparison procedure: MGR test
MRtest(y = res, trt = "treat", alpha = 0.05,
       main = "Multiple Comparison Procedure: MGR test",
       MCP = c("MGR"))

# Assuming that the available data are the averages
#  of the treatments and the analysis of variance

# Analysis of Variance Table

# Response: rv
#            Df Sum Sq Mean Sq F value    Pr(>F)
# treat      4 4135.2 1033.80  14.669 4.562e-05 ***
# Residuals 15 1057.1   70.47

mean.treat <- c(100.87, 105.95, 117.62, 127.97, 140.30)
treat <- factor(LETTERS[1:5])
DFerror <- 15
MSerror <- 70.47488
replic <- 4

MRtest(y = mean.treat,
       trt = treat,
       dferror = DFerror,
       mserror = MSerror,
       replication = replic,
       alpha = 0.05,
       main = "Multiple Comparison Procedure: MGM test",
       MCP = c("MGM"),
       ismean = TRUE)

}
